Adding pojo package and decoupling dto layer to extend pojo's
