+++ thirdeye/thirdeye-pinot/src/main/java/com/linkedin/thirdeye/datalayer/dto/MergedAnomalyResultDTO.java
@@ -1,11 +1,9 @@
 package com.linkedin.thirdeye.datalayer.dto;
 
-import com.fasterxml.jackson.annotation.JsonIgnoreProperties;
 import java.util.ArrayList;
 import java.util.List;
-import java.util.Objects;
+
 import javax.persistence.CascadeType;
-import javax.persistence.Column;
 import javax.persistence.Entity;
 import javax.persistence.FetchType;
 import javax.persistence.JoinColumn;
@@ -14,44 +12,14 @@ import javax.persistence.ManyToOne;
 import javax.persistence.OneToMany;
 import javax.persistence.OneToOne;
 import javax.persistence.Table;
-import org.apache.commons.lang.ObjectUtils;
+
+import com.fasterxml.jackson.annotation.JsonIgnoreProperties;
+import com.linkedin.thirdeye.datalayer.pojo.MergedAnomalyResultBean;
 
 @JsonIgnoreProperties(ignoreUnknown = true)
 @Entity
 @Table(name = "anomaly_merged_results")
-public class MergedAnomalyResultDTO extends AbstractDTO implements Comparable<MergedAnomalyResultDTO> {
-
-  @Column(name = "collection")
-  private String collection;
-
-  @Column(name = "metric")
-  private String metric;
-
-  @Column(name = "dimensions")
-  private String dimensions;
-
-  @Column(name = "start_time", nullable = false)
-  private Long startTime;
-
-  @Column(name = "end_time", nullable = false)
-  private Long endTime;
-
-  // significance level
-  @Column(name = "score", nullable = false)
-  private double score;
-
-  // severity
-  @Column(name = "weight", nullable = false)
-  private double weight;
-
-  @Column(name = "created_time", nullable = false)
-  private Long createdTime;
-
-  @Column(name = "message")
-  private String message;
-
-  @Column(name = "notified")
-  private boolean notified;
+public class MergedAnomalyResultDTO extends MergedAnomalyResultBean {
 
   @OneToOne(cascade = CascadeType.ALL, fetch = FetchType.EAGER, orphanRemoval = true)
   @JoinColumn(name = "anomaly_feedback_id")
@@ -66,53 +34,6 @@ public class MergedAnomalyResultDTO extends AbstractDTO implements Comparable<Me
   @JoinColumn(name = "function_id")
   private AnomalyFunctionDTO function;
 
-  public String getMetric() {
-    return metric;
-  }
-
-  public void setMetric(String metric) {
-    this.metric = metric;
-  }
-
-  public Long getStartTime() {
-    return startTime;
-  }
-
-  public void setStartTime(Long startTime) {
-    this.startTime = startTime;
-  }
-
-  public Long getEndTime() {
-    return endTime;
-  }
-
-  public void setEndTime(Long endTime) {
-    this.endTime = endTime;
-  }
-
-  public String getDimensions() {
-    return dimensions;
-  }
-
-  public void setDimensions(String dimensions) {
-    this.dimensions = dimensions;
-  }
-
-  public double getScore() {
-    return score;
-  }
-
-  public void setScore(double score) {
-    this.score = score;
-  }
-
-  public Long getCreatedTime() {
-    return createdTime;
-  }
-
-  public void setCreatedTime(Long createdTime) {
-    this.createdTime = createdTime;
-  }
 
   public AnomalyFeedbackDTO getFeedback() {
     return feedback;
@@ -130,21 +51,6 @@ public class MergedAnomalyResultDTO extends AbstractDTO implements Comparable<Me
     this.anomalyResults = anomalyResults;
   }
 
-  public boolean isNotified() {
-    return notified;
-  }
-
-  public void setNotified(boolean notified) {
-    this.notified = notified;
-  }
-
-  public String getCollection() {
-    return collection;
-  }
-
-  public void setCollection(String collection) {
-    this.collection = collection;
-  }
 
   public AnomalyFunctionDTO getFunction() {
     return function;
@@ -154,54 +60,4 @@ public class MergedAnomalyResultDTO extends AbstractDTO implements Comparable<Me
     this.function = function;
   }
 
-  public double getWeight() {
-    return weight;
-  }
-
-  public void setWeight(double weight) {
-    this.weight = weight;
-  }
-
-  public String getMessage() {
-    return message;
-  }
-
-  public void setMessage(String message) {
-    this.message = message;
-  }
-
-  @Override
-  public int hashCode() {
-    return Objects.hash(getId(), startTime, endTime, collection, metric, dimensions, score);
-  }
-
-  @Override
-  public boolean equals (Object o) {
-      if (!(o instanceof MergedAnomalyResultDTO)) {
-        return false;
-      }
-    MergedAnomalyResultDTO m = (MergedAnomalyResultDTO) o;
-    return Objects.equals(getId(), m.getId()) && Objects.equals(startTime, m.getStartTime())
-        && Objects.equals(endTime, m.getEndTime()) && Objects.equals(collection, m.getCollection())
-        && Objects.equals(metric, m.getMetric()) && Objects.equals(dimensions, m.getDimensions());
-  }
-
-  @Override
-  public int compareTo(MergedAnomalyResultDTO o) {
-    // compare by dimension, -startTime, functionId, id
-    int diff = ObjectUtils.compare(getDimensions(), o.getDimensions());
-    if (diff != 0) {
-      return diff;
-    }
-    diff = -ObjectUtils.compare(startTime, o.getStartTime()); // inverted to sort by
-    // decreasing time
-    if (diff != 0) {
-      return diff;
-    }
-    diff = ObjectUtils.compare(getFunction().getId(), o.getFunction().getId());
-    if (diff != 0) {
-      return diff;
-    }
-    return ObjectUtils.compare(getId(), o.getId());
-  }
 }
