+++ thirdeye/thirdeye-pinot/src/main/java/com/linkedin/thirdeye/datalayer/pojo/MergedAnomalyResultBean.java
@@ -1,0 +1,168 @@
+package com.linkedin.thirdeye.datalayer.pojo;
+
+import com.fasterxml.jackson.annotation.JsonIgnoreProperties;
+import java.util.ArrayList;
+import java.util.List;
+import java.util.Objects;
+import javax.persistence.CascadeType;
+import javax.persistence.Column;
+import javax.persistence.Entity;
+import javax.persistence.FetchType;
+import javax.persistence.JoinColumn;
+import javax.persistence.JoinTable;
+import javax.persistence.ManyToOne;
+import javax.persistence.OneToMany;
+import javax.persistence.OneToOne;
+import javax.persistence.Table;
+import org.apache.commons.lang.ObjectUtils;
+
+@JsonIgnoreProperties(ignoreUnknown = true)
+@Entity
+@Table(name = "anomaly_merged_results")
+public class MergedAnomalyResultBean extends AbstractBean
+    implements Comparable<MergedAnomalyResultBean> {
+
+  @Column(name = "collection")
+  private String collection;
+
+  @Column(name = "metric")
+  private String metric;
+
+  @Column(name = "dimensions")
+  private String dimensions;
+
+  @Column(name = "start_time", nullable = false)
+  private Long startTime;
+
+  @Column(name = "end_time", nullable = false)
+  private Long endTime;
+
+  // significance level
+  @Column(name = "score", nullable = false)
+  private double score;
+
+  // severity
+  @Column(name = "weight", nullable = false)
+  private double weight;
+
+  @Column(name = "created_time", nullable = false)
+  private Long createdTime;
+
+  @Column(name = "message")
+  private String message;
+
+  @Column(name = "notified")
+  private boolean notified;
+
+  public String getMetric() {
+    return metric;
+  }
+
+  public void setMetric(String metric) {
+    this.metric = metric;
+  }
+
+  public Long getStartTime() {
+    return startTime;
+  }
+
+  public void setStartTime(Long startTime) {
+    this.startTime = startTime;
+  }
+
+  public Long getEndTime() {
+    return endTime;
+  }
+
+  public void setEndTime(Long endTime) {
+    this.endTime = endTime;
+  }
+
+  public String getDimensions() {
+    return dimensions;
+  }
+
+  public void setDimensions(String dimensions) {
+    this.dimensions = dimensions;
+  }
+
+  public double getScore() {
+    return score;
+  }
+
+  public void setScore(double score) {
+    this.score = score;
+  }
+
+  public Long getCreatedTime() {
+    return createdTime;
+  }
+
+  public void setCreatedTime(Long createdTime) {
+    this.createdTime = createdTime;
+  }
+
+  public boolean isNotified() {
+    return notified;
+  }
+
+  public void setNotified(boolean notified) {
+    this.notified = notified;
+  }
+
+  public String getCollection() {
+    return collection;
+  }
+
+  public void setCollection(String collection) {
+    this.collection = collection;
+  }
+
+
+  public double getWeight() {
+    return weight;
+  }
+
+  public void setWeight(double weight) {
+    this.weight = weight;
+  }
+
+  public String getMessage() {
+    return message;
+  }
+
+  public void setMessage(String message) {
+    this.message = message;
+  }
+
+  @Override
+  public int hashCode() {
+    return Objects.hash(getId(), startTime, endTime, collection, metric, dimensions, score);
+  }
+
+  @Override
+  public boolean equals(Object o) {
+    if (!(o instanceof MergedAnomalyResultBean)) {
+      return false;
+    }
+    MergedAnomalyResultBean m = (MergedAnomalyResultBean) o;
+    return Objects.equals(getId(), m.getId()) && Objects.equals(startTime, m.getStartTime())
+        && Objects.equals(endTime, m.getEndTime()) && Objects.equals(collection, m.getCollection())
+        && Objects.equals(metric, m.getMetric()) && Objects.equals(dimensions, m.getDimensions());
+  }
+
+  @Override
+  public int compareTo(MergedAnomalyResultBean o) {
+    // compare by dimension, -startTime, functionId, id
+    int diff = ObjectUtils.compare(getDimensions(), o.getDimensions());
+    if (diff != 0) {
+      return diff;
+    }
+    diff = -ObjectUtils.compare(startTime, o.getStartTime()); // inverted to sort by
+    // decreasing time
+    if (diff != 0) {
+      return diff;
+    }
+    return ObjectUtils.compare(getId(), o.getId());
+  }
+}
