+++ thirdeye/thirdeye-pinot/src/test/java/com/linkedin/thirdeye/datalayer/ScriptRunner.java
@@ -5,12 +5,7 @@ import java.io.LineNumberReader;
 import java.io.PrintWriter;
 import java.io.Reader;
 import java.sql.Connection;
-import java.sql.ResultSet;
-import java.sql.ResultSetMetaData;
 import java.sql.SQLException;
-import java.sql.Statement;
-import java.util.regex.Matcher;
-import java.util.regex.Pattern;
 
 /**
  * Tool to run database scripts
